requirements.txt
pdf2image==1.16.3
Pillow==10.0.0
argparse==1.4.0
